load_balancer: Allow for on-demand computation.
The load balancer context is specified as a pure virtual class, but
the return type of the hashKey() member implies that there is a value
stored inside the implementation, since a const reference is returned.
Change the return type to a non-const non-reference, allowing a
stateless implementation.  Enhance the one call site that called
hashKey() twice to call it only once to avoid overhead in case each
hashKey() call recomputes the hash.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>