connection: Add support for setting SO_MARK on created client connections.
Socket mark (SO_MARK) is a 32-bit unsigned integer value that is set
on each kernel internal packet sent out from a socket, and it defaults
to 0.  This is useful for integration with other networking subsystems
in Linux, e.g., with iptables.  For example, iptables rules can be set
to match on the packet mark values to differentiate the processing the
packets with different mark values get.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>