listen_socket: Allow socket options to be set before bind().
So far the listen socket options have been set in the internal
`setSocket()` function after the socket has already been bound. This
has the benefit of (re)setting the options whenever a listener
configuration is changed, which may involve a change in listener
filter configurarion. Some socket options, however, need to be set
before the bind() call for them to have the desired effect.

This patch adds a boolean parameter to the setOptions() callback,
telling the options implementation whether the socket has already been
bound or not. The setOptions() callback is called on the listen socket
once with `bound=false` when the socket is first created right before
the bind call. Then, when whenever the listen socket is "set" the
setOptions() callback is called as before, but now typicallyl with
`bound=true`. This latter call will be done again whenever the same
socket is re-used for a new or modified listener.

For non-bound listen sockets, the setOptions() is always called with
`bound=false`.

Client connections will have the setOptions() called before the bind
(like before) and with `bound=false`. Client sockets are not recycled
like the listen sockets, so they will not have the setOptions() called
again after the bind() call.

It should be noted that setting any options via the setOptions()
callback functionality discussed here still requires the options
object be set by a listener filter. While it would be possible to
create a listener filter for the sole purpose of setting socket
options, no such filter exist yet.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>